Move all penguin entity and framework code into more appropriately named namespaces
